requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.9.3
blessed==1.19.1
botocore==1.19.60
cached-property==1.5.2
cachetools==5.3.1
cement==2.8.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
charset-normalizer==3.0.1
colorama==0.4.3
contourpy==1.0.7
cryptography==3.3.1
cycler==0.11.0
datetime-truncate==1.1.1
decorator==5.1.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.2.4
django-admin==2.0.2
django-admin-charts==1.3.0
django-allauth==0.44.0
django-appconf==1.0.4
django-bootstrap4==2.3.1
django-bower==5.2.0
django-db-logger==0.1.12
django-excel-response2==3.0.5
django-imagekit==4.0.2
django-import-export==2.6.1
django-jazzmin==2.6.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-pandas==0.6.4
django-picklefield==3.1
django-q==1.3.9
django-six==1.0.5
docopt==0.6.2
docutils==0.15.2
et-xmlfile==1.1.0
excel-base==1.0.4
executing==1.2.0
factory-boy==3.2.1
Faker==15.3.4
fonttools==4.39.4
frozenlist==1.3.3
future==0.16.0
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==2.10
importlib-metadata==2.1.1
ipython==8.13.2
isoweek==1.3.3
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
kiwisolver==1.4.4
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.21.2
oauthlib==3.1.0
odfpy==1.4.1
openai==0.27.0
openpyxl==3.0.10
packaging==23.1
pandas==1.3.3
parso==0.8.3
pathspec==0.5.9
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==8.1.0
plotly==5.3.1
prompt-toolkit==3.0.36
protobuf==4.24.4
psutil==5.9.5
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycparser==2.20
Pygments==2.13.0
PyJWT==2.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.7
screen==1.0.1
semantic-version==2.5.0
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.1
sqlparse==0.4.1
stack-data==0.6.2
stripe==8.7.0
tablib==3.3.0
tenacity==8.1.0
termcolor==1.1.0
texttable==1.6.7
TimeConvert==3.0.13
tqdm==4.64.1
traitlets==5.8.0
typing_extensions==4.10.0
tzlocal==5.0.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.59.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
zipp==0.5.0
